+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # What's new:
-- Fixed a rendering bug with bells
-
-- Added compatibility with Woodworks 
+- Bug fixes
+- Added compat with all chests using the Blueprint mod (Woodworks and more)
 
 ## If you're using Embeddium alongside OBE, you'll also need the [Forgified Fabric API](https://modrinth.com/mod/forgified-fabric-api)